requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.34
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.9.0
psutil==6.0.0
